+++ src/core_modules/capture-core-utils/rulesEngine/processors/rulesEffectsProcessor/rulesEffectsProcessor.js
@@ -171,6 +171,7 @@ export function getRulesEffectsProcessor(
             displayText: {
                 id: effect.id,
                 message: `${effect.content} ${sanitiseFalsy(effect.data)}`,
+                ...effect.style,
             },
         };
     }
@@ -183,6 +184,7 @@ export function getRulesEffectsProcessor(
                 id: effect.id,
                 key: effect.content,
                 value: effect.data,
+                ...effect.style,
             },
         };
     }