+++ src/core_modules/capture-core/components/Widget/widgetNonCollapsible.types.js
@@ -1,0 +1,14 @@
+// @flow
+import type { Node } from 'react';
+
+export type WidgetNonCollapsibleProps = {|
+    header: Node,
+    children: Node,
+    color?: string,
+    borderless?: boolean,
+|};
+
+export type WidgetNonCollapsiblePropsPlain = {|
+    ...WidgetNonCollapsibleProps,
+    ...CssClasses,
+|};