+++ src/core_modules/capture-core/components/WidgetErrorAndWarning/WidgetWarning/index.js
@@ -1,0 +1,1 @@
+export { WidgetWarning } from './WidgetWarning.component';