+++ src/core_modules/capture-core/components/WidgetFeedback/WidgetFeedback.component.js
@@ -1,23 +1,29 @@
 // @flow
-import React, { useCallback, useState } from 'react';
+import React, { useState } from 'react';
 import i18n from '@dhis2/d2-i18n';
 import { Widget } from '../Widget';
 import type { Props } from './WidgetFeedback.types';
+import { WidgetFeedbackContent } from './WidgetFeedbackContent/WidgetFeedbackContent';
 
-export const WidgetFeedback = ({ className }: Props) => {
-    const [openStatus, setOpenStatus] = useState(false);
+export const WidgetFeedback = ({ feedback, emptyText }: Props) => {
+    const [openStatus, setOpenStatus] = useState(true);
+
     return (
         <div
-            className={className}
+            data-test="feedback-widget"
         >
             <Widget
                 header={i18n.t('Feedback')}
-                onOpen={useCallback(() => setOpenStatus(true), [setOpenStatus])}
-                onClose={useCallback(() => setOpenStatus(false), [setOpenStatus])}
+                onOpen={() => setOpenStatus(true)}
+                onClose={() => setOpenStatus(false)}
                 open={openStatus}
             >
-                <p>Feedback Placeholder Text</p>
+                <WidgetFeedbackContent
+                    widgetData={feedback}
+                    emptyText={emptyText}
+                />
             </Widget>
         </div>
     );
 };
+