+++ src/core_modules/capture-core/metaDataMemoryStoreBuilders/programs/factory/enrollment/EnrollmentFactory.js
@@ -5,26 +5,20 @@ import i18n from '@dhis2/d2-i18n';
 import { capitalizeFirstLetter } from 'capture-core-utils/string/capitalizeFirstLetter';
 import { errorCreator } from 'capture-core-utils';
 import type {
+    CachedDataEntryForm,
     CachedProgram,
+    CachedProgramSection,
     CachedProgramTrackedEntityAttribute,
     CachedTrackedEntityAttribute,
     CachedTrackedEntityType,
 } from '../../../../storageControllers/cache.types';
-import {
-    RenderFoundation,
-    Section,
-    Enrollment,
-    CustomForm,
-    InputSearchGroup,
-} from '../../../../metaData';
-import type {
-    TrackedEntityType,
-    SearchGroup,
-} from '../../../../metaData';
+import type { SearchGroup, TrackedEntityType } from '../../../../metaData';
+import { CustomForm, Enrollment, InputSearchGroup, RenderFoundation, Section } from '../../../../metaData';
 import { DataElementFactory } from './DataElementFactory';
 import { getApi } from '../../../../d2/d2Instance';
 import { DataElement } from '../../../../metaData/DataElement';
 import type { ConstructorInput } from './enrollmentFactory.types';
+import { transformTrackerNode } from '../transformNodeFuntions/transformNodeFunctions';
 
 export class EnrollmentFactory {
     static errorMessages = {
@@ -91,63 +85,137 @@ export class EnrollmentFactory {
         return DataElementFactory.buildtetFeatureType(featureType);
     }
 
-    async _buildSection(
-        cachedProgramTrackedEntityAttributes: ?Array<CachedProgramTrackedEntityAttribute>,
-        cachedProgramTrackedEntityTypeId: ?string,
+    async _buildTetFeatureTypeSection(
+        cachedProgramTrackedEntityTypeId: string,
     ) {
         const featureTypeField = this._buildTetFeatureTypeField(cachedProgramTrackedEntityTypeId);
-        if ((!cachedProgramTrackedEntityAttributes ||
-            cachedProgramTrackedEntityAttributes.length <= 0) &&
-            !featureTypeField) {
+        const trackedEntityType = this.cachedTrackedEntityTypes.get(cachedProgramTrackedEntityTypeId);
+
+        if (!featureTypeField) {
             return null;
         }
 
+        const section = new Section((o) => {
+            o.id = cachedProgramTrackedEntityTypeId;
+            o.name = trackedEntityType?.displayName || '';
+        });
+
+        featureTypeField && section.addElement(featureTypeField);
+        return section;
+    }
+
+    async _buildMainSection(
+        cachedProgramTrackedEntityAttributes?: ?Array<CachedProgramTrackedEntityAttribute>,
+        cachedProgramTrackedEntityTypeId?: ?string,
+    ) {
         const section = new Section((o) => {
             o.id = Section.MAIN_SECTION_ID;
             o.name = i18n.t('Profile');
         });
 
-        featureTypeField && section.addElement(featureTypeField);
-        if (cachedProgramTrackedEntityAttributes && cachedProgramTrackedEntityAttributes.length > 0) {
-            // $FlowFixMe
-            await cachedProgramTrackedEntityAttributes.asyncForEach(async (ptea) => {
-                const element = await this.dataElementFactory.build(ptea);
-                element && section.addElement(element);
-            });
+        if (!cachedProgramTrackedEntityAttributes?.length) { return null; }
+
+        if (cachedProgramTrackedEntityTypeId) {
+            const featureTypeField = this._buildTetFeatureTypeField(cachedProgramTrackedEntityTypeId);
+            featureTypeField && section.addElement(featureTypeField);
         }
 
+        await this._buildElementsForSection(cachedProgramTrackedEntityAttributes, section);
         return section;
+    }
+
+    async _buildElementsForSection(
+        cachedProgramTrackedEntityAttributes: ?Array<CachedProgramTrackedEntityAttribute>,
+        section: Section,
+    ) {
+        // $FlowFixMe
+        await cachedProgramTrackedEntityAttributes.asyncForEach(async (trackedEntityAttribute) => {
+            const element = await this.dataElementFactory.build(trackedEntityAttribute);
+            element && section.addElement(element);
+        });
+        return section;
+    }
+
+    async _buildSection(
+        cachedProgramTrackedEntityAttributes?: Array<CachedProgramTrackedEntityAttribute>,
+        cachedSectionCustomLabel: string,
+        cachedSectionCustomId: string,
+    ) {
+        if (!cachedProgramTrackedEntityAttributes?.length) {
+            return null;
+        }
+
+        const section = new Section((o) => {
+            o.id = cachedSectionCustomId;
+            o.name = cachedSectionCustomLabel;
+        });
+
+        await this._buildElementsForSection(cachedProgramTrackedEntityAttributes, section);
+        return section;
+    }
+
+    async _buildCustomEnrollmentForm(
+        enrollmentForm: RenderFoundation,
+        dataEntryForm: CachedDataEntryForm,
+        cachedProgramTrackedEntityAttributes: ?Array<CachedProgramTrackedEntityAttribute>,
+    ) {
+        if (!cachedProgramTrackedEntityAttributes) { return null; }
+
+        let section = new Section((o) => {
+            o.id = Section.MAIN_SECTION_ID;
+        });
+
+        section.showContainer = false;
+
+        section = await this._buildElementsForSection(cachedProgramTrackedEntityAttributes, section);
+        section && enrollmentForm.addSection(section);
+        try {
+            section.customForm = new CustomForm((o) => {
+                o.id = dataEntryForm.id;
+            });
+            section.customForm.setData(dataEntryForm.htmlCode, transformTrackerNode);
+        } catch (error) {
+            log.error(errorCreator(EnrollmentFactory.errorMessages.CUSTOM_FORM_TEMPLATE_ERROR)({
+                template: dataEntryForm.htmlCode, error, method: 'buildEnrollment' }));
+        }
+        return enrollmentForm;
     }
 
     async _buildEnrollmentForm(
         cachedProgram: CachedProgram,
+        cachedProgramSections: ?Array<CachedProgramSection>,
     ) {
+        const cachedProgramTrackedEntityAttributes = cachedProgram?.programTrackedEntityAttributes;
         const enrollmentForm = new RenderFoundation((o) => {
             o.featureType = EnrollmentFactory._getFeatureType(cachedProgram.featureType);
             o.name = cachedProgram.displayName;
         });
 
-        let section =
-            await this._buildSection(cachedProgram.programTrackedEntityAttributes, cachedProgram.trackedEntityTypeId);
-        section && enrollmentForm.addSection(section);
-
+        let section;
         if (cachedProgram.dataEntryForm) {
-            if (!section) {
-                section = new Section((o) => {
-                    o.id = Section.MAIN_SECTION_ID;
-                });
+            if (cachedProgram.trackedEntityTypeId) {
+                section = await this._buildTetFeatureTypeSection(cachedProgram.trackedEntityTypeId);
+                section && enrollmentForm.addSection(section);
             }
-            section.showContainer = false;
-            const dataEntryForm = cachedProgram.dataEntryForm;
-            try {
-                enrollmentForm.customForm = new CustomForm((o) => {
-                    o.id = dataEntryForm.id;
-                    o.data = dataEntryForm.htmlCode;
-                });
-            } catch (error) {
-                log.error(errorCreator(EnrollmentFactory.errorMessages.CUSTOM_FORM_TEMPLATE_ERROR)({
-                    template: dataEntryForm.htmlCode, error, method: 'buildEnrollment' }));
+
+            await this._buildCustomEnrollmentForm(enrollmentForm, cachedProgram.dataEntryForm, cachedProgramTrackedEntityAttributes);
+        } else if (cachedProgramSections?.length) {
+            if (cachedProgram.trackedEntityTypeId) {
+                section = await this._buildTetFeatureTypeSection(cachedProgram.trackedEntityTypeId);
+                section && enrollmentForm.addSection(section);
             }
+
+            // $FlowFixMe
+            cachedProgramTrackedEntityAttributes && cachedProgramSections.asyncForEach(async (programSection) => {
+                const trackedEntityAttributes = cachedProgramTrackedEntityAttributes
+                    .filter(trackedEntityAttribute => programSection.trackedEntityAttributes
+                        .includes(trackedEntityAttribute.trackedEntityAttributeId));
+                section = await this._buildSection(trackedEntityAttributes, programSection.displayFormName, programSection.id);
+                section && enrollmentForm.addSection(section);
+            });
+        } else {
+            section = await this._buildMainSection(cachedProgramTrackedEntityAttributes, cachedProgram.trackedEntityTypeId);
+            section && enrollmentForm.addSection(section);
         }
         return enrollmentForm;
     }
@@ -260,7 +328,7 @@ export class EnrollmentFactory {
             o.inputSearchGroups = this._buildInputSearchGroups(cachedProgram, programSearchGroups);
         });
 
-        enrollment.enrollmentForm = await this._buildEnrollmentForm(cachedProgram);
+        enrollment.enrollmentForm = await this._buildEnrollmentForm(cachedProgram, cachedProgram.programSections);
         return enrollment;
     }
 }